refactor: move InMemoryStatusRepository to infrastructure
